% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{build_binomial}
\alias{build_binomial}
\title{Helper functions for binomial models}
\usage{
build_binomial(y)
}
\arguments{
\item{y}{A two-levels factor or a vector containing only 0 and 1 values.}
}
\value{
A vector containing only 0 and 1 values.
}
\description{
Helper functions for binomial models
}
