% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_models.R
\name{Exp}
\alias{Exp}
\title{Create an Exponential Function and its Derivatives}
\usage{
Exp()
}
\value{
A list containing three functions:
\item{f}{The exponential function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates
\item \code{theta}: A vector of coefficients
}
}
\item{J}{The Jacobian (first derivatives) function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates
\item \code{theta}: A vector of coefficients
\item \code{y}: Optional; pre-computed output values from the exponential function
}
}
\item{H}{The Hessian (second derivatives) function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates
\item \code{theta}: A vector of coefficients
\item \code{y}: Optional; pre-computed output values from the exponential function
}
Returns a list of Hessian matrices, one for each observation in \code{X}
}
}
\description{
This function creates an exponential function and its derivatives (Jacobian and Hessian) for use in optimization, modeling, and statistical analysis.
}
\details{
The exponential function implemented here has the following form:
\deqn{f(x_i, \theta) = \exp(x_i^\intercal \theta)}

Where:
\itemize{
\item \eqn{x_i^{\intercal}} is the \eqn{i}-th row vector of the matrix of covariates \eqn{X}, with each row representing an observation
and each column representing a covariate
\item \eqn{\theta} is a vector of coefficients
\item The function returns a vector of exponential values, one for each row of \eqn{X}
}

The Jacobian matrix has the form:
\deqn{J_{ij} = \dfrac{\partial y_i}{\partial \theta_j} =  y_i x_{ij}}

Where:
\itemize{
\item \eqn{y_i} is the output from the exponential function for observation \eqn{i}
\item \eqn{x_{ij}} is the value of covariate \eqn{j} for observation \eqn{i}
}

The Hessian matrix for the generic \eqn{i}-th observation has the sctructure:
\deqn{H_{jl} = H_{lj} = \dfrac{\partial^2y_i}{\partial \theta_j \partial \theta_l} = y_i x_{ij} x_{il}}

Where:
\itemize{
\item \eqn{y_i} is the output from the exponential function for the \eqn{i}-th observation
\item \eqn{x_{ij}} and \eqn{x_{il}} are values of covariates \eqn{j} and \eqn{l} for the \eqn{i}-th observation
}
}
\examples{
# Create the exponential function
exp_func <- Exp()

# Define covariate data
X <- matrix(c(1, 0.5, 0.2,
              1, 0.8, 0.3,
              1, 0.6, 0.4), ncol = 3, byrow = TRUE)

# Define coefficients
theta <- c(0.1, 0.5, -0.3)

# Calculate exponential values
y_values <- exp_func$f(X, theta)
y_values

# Calculate Jacobian (first derivatives)
jacobian <- exp_func$J(X, theta)
jacobian

# Calculate Hessian matrices (second derivatives)
hessian_list <- exp_func$H(X, theta)
hessian_list[[1]]  # Hessian matrix for the first observation

}
