% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnlmsa.R
\name{gnlmsa}
\alias{gnlmsa}
\title{Fitting Generalized Non-Linear Models}
\usage{
gnlmsa(
  y,
  X,
  Z,
  family,
  f_mu,
  J_mu,
  H_mu,
  f_phi,
  J_phi,
  H_phi,
  beta_start,
  lower_mu,
  upper_mu,
  gamma_start,
  lower_phi,
  upper_phi,
  mult,
  nsim,
  sa_control = sa_control(),
  maxit = 100,
  tol = 1e-05,
  expected = TRUE,
  verbose = TRUE,
  beta_names,
  gamma_names
)
}
\arguments{
\item{y}{numerical vector of response values.}

\item{X}{design matrix for the mean model.}

\item{Z}{design matrix for the dispersion model.}

\item{family}{a \code{\link{family_gnlmsa}} object.}

\item{f_mu}{predictor function for the mean component.}

\item{J_mu}{(optional) Jacobian of \code{f_mu} with respect to the parameters.}

\item{H_mu}{(optional) Hessian of \code{f_mu} with respect to the parameters.}

\item{f_phi}{predictor function for the dispersion component.}

\item{J_phi}{(optional) Jacobian of \code{f_phi} with respect to the parameters.}

\item{H_phi}{(optional) Hessian of \code{f_phi} with respect to the parameters.}

\item{beta_start}{initial values for the parameters of the mean component.}

\item{lower_mu}{lower bounds for the parameters of the mean component.}

\item{upper_mu}{upper bounds for the parameters of the mean component.}

\item{gamma_start}{initial values for the parameters of the dispersion component.}

\item{lower_phi}{lower bounds for the parameters of the dispersion component.}

\item{upper_phi}{upper bounds for the parameters of the dispersion component.}

\item{mult}{numerical vector multipliers of errors in Simulated Annealing algorithm.}

\item{nsim}{number of Simulated Annealing simulations to perform (possibly in parallel).}

\item{sa_control}{list of control parameters for Simulated Annealing algorithm, created with \code{sa_control()}}

\item{maxit}{integer. Maximum number of iterations in Newton-Raphson optimization.}

\item{tol}{convergence tolerance on the parameters (default 1e‑5).}

\item{expected}{logical indicating whether to use the expected (default, \code{expected = TRUE}) or the observed (\code{expected = FALSE}) in the optimization algorithm.}

\item{verbose}{logical indicating whether to print progress information.}

\item{beta_names}{(optional) character vector containing the names of parameters of dispersion component. Must be of the same length of \code{gamma_start}, \code{lower_phi} and \code{upper_phi}.}
}
\value{
A list.
}
\description{
Fitting Generalized Non-Linear Models
}
\examples{
\dontrun{
y <- productivity$Y
X <- as.matrix(productivity[, -1])
Z <- cbind(1, prcomp(scale(X))$x[, 1])
family <- gnlmsa_Gamma("identity")
f_mu <- cobb_douglas()$f
J_mu <- cobb_douglas()$J
H_mu <- cobb_douglas()$H
f_phi <- Linear()$f
J_phi <- Linear()$J
H_phi <- Linear()$H
beta_start <- c(10, .5, .5)
lower_mu <- rep(0, 3)
upper_mu <- rep(Inf, 3)
gamma_start <- c(log(var(y)), 0)
lower_phi <- rep(-Inf, 2)
upper_phi <- rep(Inf, 2)
mult <- 1
nsim <- 1
sa_control <- sa_control(10000)
maxit <- 100
tol <- 1e-05
expected <- TRUE
verbose <- TRUE
beta_names <- c("a", "alpha", "beta")
gnlmsa(y, X, Z, family,
       f_mu, J_mu, H_mu,
       f_phi, J_phi, H_phi,
       beta_start, lower_mu, upper_mu,
       gamma_start, lower_phi, upper_phi,
       mult, nsim, sa_control = sa_control(),
       maxit = 100, tol = 1e-05,
       expected = TRUE, verbose = TRUE,
       beta_names)
}
}
