% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{gnlmsa_poisson}
\alias{gnlmsa_poisson}
\title{Poisson Family for Generalized Non-Linear Models}
\usage{
gnlmsa_poisson(link_mu = "log")
}
\arguments{
\item{link_mu}{A character string specifying the link function for the mean component \eqn{\mu}.
Must be one of \code{"log"} (default), \code{"identity"} or \code{"sqrt"}.}
}
\value{
An object of class \code{family_gnlmsa} (inheriting from \code{family}) with the following components:
\describe{
\item{family}{Name of the family: \code{"gnlmsa_poisson"}}

\item{link_mu}{Name of the link function used for the mean.}
\item{linkfun_mu}{Function mapping \eqn{\mu} to the predictor \eqn{\eta}.}
\item{linkinv_mu}{Inverse of the link function, mapping \eqn{\eta} to \eqn{\mu}.}
\item{mu.eta}{First derivative of \eqn{\mu} with respect to \eqn{\eta}.}
\item{mu2.eta2}{Second derivative of \eqn{\mu} with respect to \eqn{\eta}.}
\item{validmu}{Validator function for the mean: checks positivity and finiteness.}

\item{link_phi}{Name of the link function used for the dispersion.}
\item{linkfun_phi}{Function mapping \eqn{\phi} to the predictor \eqn{v}.}
\item{linkinv_phi}{Inverse of the link function, mapping \eqn{v} to \eqn{\phi}.}
\item{phi.vi}{First derivative of \eqn{\phi} with respect to \eqn{v}.}
\item{phi2.vi2}{Second derivative of \eqn{\phi} with respect to \eqn{v}.}
\item{validphi}{Validator function for the dispersion: checks positivity and finiteness.}

\item{variance}{Function defining the conditional variance: \eqn{Var(Y_i \mid \mu_i, \phi_i) = \mu_i}}
\item{loglik}{Log-likelihood contribution function for each observation.}
\item{grad_mu}{Gradient of the log-likelihood with respect to the mean parameters.}
\item{hess_mu}{Hessian of the log-likelihood with respect to the mean parameters.}
\item{grad_phi}{Gradient of the log-likelihood with respect to the dispersion parameters.}
\item{hess_phi}{Hessian of the log-likelihood with respect to the dispersion parameters.}
\item{hess_mu_phi}{Cross-derivative of the log-likelihood with respect to mean and dispersion parameters.}

\item{simulate}{Function to simulate to be called by method \code{simulate}.}
}
}
\description{
Constructs a \code{family_gnlmsa} object for use with Poisson-distributed response variables in
Generalized Non-Linear Models (GNLMs). In the Poisson the \eqn{\phi} parameter is fixed to be 1.
}
\details{
The Poisson distribution is commonly used to model count data.
The pdf of the Poisson distribution is:
\deqn{
f(y_i; \mu_i) =
\dfrac{\mu_i^{y_i} e^{\mu_i}}{y_i!}
}

The log-likelihood function (summed over all observations) is:
\deqn{
\ell(\mu) = \sum_{i=1}^n
y_i \log \mu_i - \mu_i - \log y_i!
}
}
\seealso{
\code{\link[=family_gnlmsa]{family_gnlmsa()}}, \code{\link[=gnlmsa]{gnlmsa()}}.
}
