% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{grad_mu}
\alias{grad_mu}
\title{Compute gradient with respect to parameters of mean component for Generalized Non-Linear Models}
\usage{
grad_mu(y, X, beta, mu, eta, phi, f_mu, J_mu, mu.eta, variance)
}
\arguments{
\item{y}{numerical vector of response variable.}

\item{X}{matrix containing the covariates for the mean component.}

\item{beta}{parameters of the mean component.}

\item{mu}{numerical vector containing the conditional mean of the response variable.}

\item{eta}{numerical vector containing the non linear predictor of the response variable.}

\item{phi}{numerical vector containing the conditional dispersion of the response variable.}

\item{f_mu}{function with signature \code{f(X, theta)} to compute the non linear predictor where \code{X} is the matrix containing the covariates and \code{theta} is the vector of parameters both of the mean component.}

\item{J_mu}{(optional) function to compute the jacobian of \code{f_mu()}. Should be a function with signature \code{J_mu(X, theta)} and should return a \eqn{n \times k} matrix where \code{n} is equal to \code{length(y)} and \code{k} is equal to the number of parameters of the mean component.}

\item{mu.eta}{function which computes the derivative of the mean component with respect to the non linear predictor as created by \code{\link[=make_link]{make_link()}} function.}

\item{variance}{function which computes the variance of the given family as a function of \code{mu} and \code{phi}.}
}
\value{
a numerical vector of length equal to the number of parameters for the mean component.
}
\description{
Compute gradient with respect to parameters of mean component for Generalized Non-Linear Models
}
\examples{
# create family object
fam_gamma <- gnlmsa_Gamma(link_mu = "identity", link_phi = "log")

# create non linear function for mean
cd <- cobb_douglas()
f_mu <- cd$f
J_mu <- cd$J

# create non linear function for dispersion
lin <- Linear()
f_phi <- lin$f
J_phi <- lin$J

# sample size
n <- 100

# covariates for the mean component
X <- cbind(rgamma(n, 1000, 10), rgamma(n, 2000, 5))

# covariates for the variance component
Z <- cbind(1, X)

# parameters for mean component
beta <- c(1, .5, .3)
eta <- f_mu(X, beta)
mu <- fam_gamma$linkinv_mu(eta)

# parameters for dispersion component
gamma <- c(1, .2, -.05)
vi <- f_phi(Z, gamma)
phi <- fam_gamma$linkinv_phi(vi)

# simulate response
y <- rgamma(n, phi, phi/mu)

grad_mu(y, X, beta, mu, eta, phi,
        f_mu, J_mu,
        fam_gamma$mu.eta, fam_gamma$variance)
}
