% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{hess}
\alias{hess}
\title{Build full hessian of a Generalized Non-Linear model}
\usage{
hess(h_mu, h_phi, h_mu_phi)
}
\arguments{
\item{h_mu}{hessian matrix for the mean component.}

\item{h_phi}{hessian matrix for the dispersion component.}

\item{h_mu_phi}{hessian matrix for the mean and dispersion component.}
}
\value{
a \eqn{(k + p) \times (k + p)} square symmetric matrix where \eqn{k} and \eqn{p} are the number of parameters for the mean component and the dispersion component respectively.
}
\description{
Build full hessian of a Generalized Non-Linear model
}
\examples{
# create family object
fam_gamma <- gnlmsa_Gamma(link_mu = "identity", link_phi = "log")

# create non linear function for mean
cd <- cobb_douglas()
f_mu <- cd$f
J_mu <- cd$J
H_mu <- cd$H

# create non linear function for dispersion
lin <- Linear()
f_phi <- lin$f
J_phi <- lin$J
H_phi <- lin$H

# sample size
n <- 100

# covariates for the mean component
X <- cbind(rgamma(n, 1000, 10), rgamma(n, 2000, 5))

# covariates for the variance component
Z <- cbind(1, X)

# parameters for mean component
beta <- c(1, .5, .3)
eta <- f_mu(X, beta)
mu <- fam_gamma$linkinv_mu(eta)

# parameters for dispersion component
gamma <- c(1, .2, -.05)
vi <- f_phi(Z, gamma)
phi <- fam_gamma$linkinv_phi(vi)

# simulate response
y <- rgamma(n, phi, phi/mu)

# compute h_mu
h_mu <- hess_mu(y, X, beta, mu, eta, phi,
                f_mu, J_mu, H_mu,
                fam_gamma$mu.eta, fam_gamma$mu2.eta2, fam_gamma$variance)

# compute h_phi
h_phi <- fam_gamma$hess_phi(y, Z, gamma,
                            phi, vi, mu,
                            f_phi, J_phi, H_phi,
                            fam_gamma$phi.vi, fam_gamma$phi2.vi2)

# compute h_mu_phi
h_mu_phi <- fam_gamma$hess_mu_phi(y, X, Z, beta, gamma,
                                  mu, eta, phi, vi,
                                  f_mu, J_mu, f_phi, J_phi,
                                  fam_gamma$mu.eta, fam_gamma$phi.vi)

# build full hessian
hess(h_mu, h_phi, h_mu_phi)
}
