% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_models.R
\name{Linear}
\alias{Linear}
\title{Create a Linear Function and its Derivatives}
\usage{
Linear(X, intercept = T, lower, upper)
}
\arguments{
\item{X}{A matrix or data frame of covariates (input data). Each row represents an observation and each column represents a covariate.}

\item{intercept}{A logical indicating whether to include an intercept term (default is TRUE).}

\item{lower}{A vector of lower bounds for the coefficients. If not specified, defaults to \code{-Inf} for all coefficients.}

\item{upper}{A vector of upper bounds for the coefficients. If not specified, defaults to \code{Inf} for all coefficients.}
}
\value{
A list containing three functions and additional information:
\item{f}{The linear function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates
\item \code{theta}: A vector of coefficients
}
}
\item{J}{The Jacobian (first derivatives) function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates
\item \code{theta}: A vector of coefficients
\item \code{y}: Optional; pre-computed output values from the linear function
}
}
\item{H}{The Hessian (second derivatives) function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates
\item \code{theta}: A vector of coefficients
\item \code{y}: Optional; pre-computed output values from the linear function
}
Returns a list of Hessian matrices, one for each observation in \code{X}
}
\item{lower}{A vector of lower bounds for the coefficients (if specified).}
\item{upper}{A vector of upper bounds for the coefficients (if specified).}
\item{X}{The input matrix or data frame of covariates.}
}
\description{
This function creates a linear function and its derivatives (Jacobian and Hessian) for use in optimization, modeling, and statistical analysis.
}
\details{
The linear function implemented here has the following form:
\deqn{f(x_i, \theta) = x_i^{\intercal}\theta}

Where:
\itemize{
\item \eqn{x_i^{\intercal}} is the \eqn{i}-th row vector of the matrix of covariates \eqn{X}, with each row representing an observation
and each column representing a covariate
\item \eqn{\theta} is a vector of coefficients
\item The function returns a vector of linear combinations, one for each row of \eqn{X}
}

The Jacobian (first derivatives) matrix has the form:
\deqn{J_{ij} = \dfrac{\partial y_i}{\partial \theta_j} =  x_{ij}}

Where:
\itemize{
\item \eqn{x_{ij}} is the value of covariate \eqn{j} for observation \eqn{i}
}

The Hessian (second derivatives) is a zero matrix of dimension \eqn{k \times k} where \eqn{k} is the number of columns of \eqn{X}.

Additionally, the function allows the inclusion of an intercept term and constraints on the coefficients (lower and upper bounds).
}
