% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_models.R
\name{Logistic}
\alias{Logistic}
\title{Create a Logistic Function and its Derivatives}
\usage{
Logistic(X, lower, upper)
}
\arguments{
\item{X}{A matrix or data frame of covariates (input data). Each row represents an
observation, and each column represents a covariate.}

\item{lower}{A vector of lower bounds for the coefficients. If not specified, defaults
to \code{-Inf} for all coefficients.}

\item{upper}{A vector of upper bounds for the coefficients. If not specified, defaults
to \code{Inf} for all coefficients.}
}
\value{
A list containing three functions and additional information:
\item{f}{The logistic function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates.
\item \code{theta}: A vector of coefficients.
}
}
\item{J}{The Jacobian (first derivatives) function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates.
\item \code{theta}: A vector of coefficients.
}
}
\item{H}{The Hessian (second derivatives) function. Takes parameters:
\itemize{
\item \code{X}: A matrix or data frame of covariates.
\item \code{theta}: A vector of coefficients.
}
Returns a list of Hessian matrices, one for each observation in \code{X}.
}
\item{lower}{A vector of lower bounds for the coefficients.}
\item{upper}{A vector of upper bounds for the coefficients.}
\item{X}{The input matrix or data frame of covariates.}
}
\description{
This function creates a logistic function and its derivatives (Jacobian and Hessian)
for use in optimization, modeling, and statistical analysis, particularly as a functional
predictor in generalized non-linear models (GNLM).
}
\details{
The logistic function implemented here has the following form:
\deqn{f(x_i, \theta) = \dfrac{\theta_1}{1 + \exp(-\theta_2(x_i - \theta_3)))}}

Where:
\itemize{
\item \eqn{x_i} is the \eqn{i}-th observation of input data \eqn{X}.
\item \eqn{\theta_1} is the scaling parameter of the logistic function.
\item \eqn{\theta_2} is the slope (rate of growth or decay).
\item \eqn{\theta_3} is the threshold value where the logistic function is centered.
\item The function returns a vector of logistic function values for each observation in \eqn{X}.
}

Additionally, the function allows the inclusion of bounds for the optimization parameters
(lower and upper bounds) and is designed to be used in optimization-based tasks like
non-linear regression or GLM fitting.
}
