% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_positive}
\alias{map_positive}
\title{Map the positive line to the real line}
\usage{
map_positive(x, lower = 0, inverse = FALSE, derivative = FALSE)
}
\arguments{
\item{x}{a numerical vector of values to map to the real line or to take back to the positive line.}

\item{lower}{value of the lower bound in order to map from \eqn{(\texttt{lower}, \infty) \rightarrow \mathbb{R}}. \code{lower} could also be a negative value. Default is 0.}

\item{inverse}{logical indicating whether to map from the positive line to the real line (default, \code{inverse = FALSE}) or whether to map from the real line to the positive line (\code{inverse = TRUE}).}

\item{derivative}{logical indicating whether to return the derivative of the map function.}
}
\value{
a numerical vector containing the mapped values or the derivative.
}
\description{
Map the positive line to the real line
}
\examples{
# (0, inf) to R
x <- 1:5
y <- map_positive(x)
map_positive(y, inverse = TRUE)

# (a, Inf) to R

x <- 2:6
lower <- 1
y <- map_positive(x, lower = lower)
map_positive(y, inverse = TRUE, lower = lower)

# derivative
x <- 1:5
map_positive(x, derivative = TRUE)
}
