% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.family_gnlmsa}
\alias{print.family_gnlmsa}
\title{Print Method for \code{family_gnlmsa} Objects}
\usage{
\method{print}{family_gnlmsa}(x, ...)
}
\arguments{
\item{x}{An object of class \code{family_gnlmsa}.}

\item{...}{Additional arguments passed to or from other methods (currently unused).}
}
\value{
The input object is returned invisibly. The function is called for its side effect of printing information to the console.
}
\description{
Displays a summary of the distribution family and associated link functions used for modeling the mean and dispersion components in a \code{gnlmsa} model.
}
\details{
This method is automatically invoked when printing a \code{family_gnlmsa} object.
}
\examples{
gamma_fam <- gnlmsa_Gamma()
gamma_fam
}
