% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{productivity}
\alias{productivity}
\title{Productivity data example for Cobb-Douglas estimation}
\format{
\subsection{\code{productivity}}{

A data frame with 60 rows and 3 columns:
\describe{
\item{Y}{value of production in dollars}
\item{X1, X2}{value of input X1 and X2 in dollars}
}
}
}
\usage{
productivity
}
\description{
Example of typical data encountered in Cobb-Douglas production function estimation.
}
\keyword{datasets}
