% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{sa_control}
\alias{sa_control}
\title{Options for Simulated Annealing algorithm}
\usage{
sa_control(
  iterations = 1000,
  compute_v = floor(iterations/10),
  initial_temperature = 100,
  final_temperature = 1,
  restart_if_stuck = floor(iterations * 0.33),
  save_history = FALSE
)
}
\arguments{
\item{iterations}{integer indicating the number of iterations to perform.}

\item{compute_v}{how many iterations wait in order to recompute variance-covariance matrix.}

\item{initial_temperature}{initial temperature of the algorithm.}

\item{final_temperature}{final temperature of the algorithm.}

\item{restart_if_stuck}{how many iterations wait before restarting from last best parameters value if the algorithm stucks.}

\item{save_history}{whether to save full history of the algorithm.}
}
\value{
a list containing the specified options.
}
\description{
Options for Simulated Annealing algorithm
}
