% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sample_par}
\alias{sample_par}
\title{Sample new value of parameters}
\usage{
sample_par(par, v, mult = 1, npar = length(par))
}
\arguments{
\item{par}{The precedent values of the parameters.}

\item{v}{The variance-covariance matrix used to sample \code{eps}.}

\item{mult}{A numerical vector to multiply \code{eps}.}

\item{npar}{A numerical vector indicating the number of parameters.}
}
\value{
A numerical vector with the new values of the parameters.
}
\description{
Sample new value of parameters
}
\examples{
par <- c(0, 2, .2)
lower <- c(-Inf, 0, -1)
upper <- c(Inf, Inf, 1)
map_functions <- make_map_function(lower, upper)
map <- map_functions$map
invert <- map_functions$invert
jacobian <- map_functions$jacobian
mapped_pars <- map(par)
j <- diag(jacobian(par))
v <- diag(1e-06, length(par))
mapped_v <- j\%*\%v\%*\%t(j)
mult <- c(3, 1, .2)
new_par_mapped <- sample_par(par = mapped_pars, v = mapped_v, mult = mult)
invert(new_par_mapped)
}
