% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnlmsa.R
\name{gnlmsa}
\alias{gnlmsa}
\title{Fit a Generalized Non-Linear Model with Mean and Dispersion Components}
\usage{
gnlmsa(
  y,
  mean_model,
  beta_start,
  family = gnlmsa_gaussian(),
  dispersion_model = NULL,
  gamma_start,
  mult = 1,
  nsim = 1,
  sa_control_params = sa_control(),
  nr_control_params = nr_control(),
  fixed_params = NULL,
  beta_names,
  gamma_names
)
}
\arguments{
\item{y}{Numeric vector of response values.}

\item{mean_model}{A list that includes the nonlinear predictor function, Jacobian, Hessian, lower and upper bounds,
and design matrix for the mean component. The list should have the following components:
\itemize{
\item \code{f}: Function for the mean component nonlinear predictor.
\item \code{J}: Jacobian of the f function.
\item \code{H}: Hessian of the f function.
\item \code{lower}: Lower bounds for the mean component parameters.
\item \code{upper}: Upper bounds for the mean component parameters.
\item \code{X}: Design matrix for the mean component.
}}

\item{beta_start}{Initial values for the parameters of the mean component.}

\item{family}{A family_gnlmsa object specifying the distributional assumptions and required derivatives.}

\item{dispersion_model}{(Optional) A list that includes the nonlinear predictor function, Jacobian, Hessian,
lower and upper bounds, and design matrix for the dispersion component. The list should have the
following components:
\itemize{
\item \code{f}: Function for the dispersion component nonlinear predictor.
\item \code{J}: Jacobian of the f function.
\item \code{H}: Hessian of the f function.
\item \code{lower}: Lower bounds for the dispersion component parameters.
\item \code{upper}: Upper bounds for the dispersion component parameters.
\item \code{X}: Design matrix for the dispersion component.
}}

\item{gamma_start}{Initial values for the parameters of the dispersion component.}

\item{mult}{Proposal scaling factor for the Simulated Annealing algorithm.}

\item{nsim}{Number of independent Simulated Annealing simulations (currently only 1 is supported).}

\item{sa_control_params}{A list of Simulated Annealing options created via \code{\link[=sa_control]{sa_control()}}.}

\item{nr_control_params}{A list of Newton-Raphson options created via \code{\link[=nr_control]{nr_control()}}.}

\item{fixed_params}{A list containing two numeric vectors of the same length:
\itemize{
\item The first vector specifies the indices of the parameters to be fixed.
\item The second vector provides the corresponding fixed values for those parameters.
}}

\item{beta_names}{(Optional) Names of the parameters in the mean component.}

\item{gamma_names}{(Optional) Names of the parameters in the dispersion component.}
}
\value{
A list of class \code{"gnlmsa"} containing:
\describe{
\item{coef}{Named vector of estimated parameters (concatenation of \code{beta} and \code{gamma}).}
\item{beta, gamma}{Estimated coefficients for mean and dispersion components.}
\item{loglik}{Log-likelihood of the final fitted model.}
\item{eta, mu}{Predictor and fitted values for the mean component.}
\item{vi, phi}{Predictor and fitted values for the dispersion component.}
\item{nsim, mult}{Settings used for Simulated Annealing.}
\item{nobs, df, df.residuals}{Sample size and degrees of freedom.}
\item{npar_mu}{Number of parameters for the mean component.}
\item{nr_failed}{Logical; \code{TRUE} if the Newton–Raphson step failed.}
\item{nr_better}{Logical; \code{TRUE} if Newton–Raphson log-likelihood was better than Simulated Annealing.}
\item{mean_model, dispersion_model}{List of functions used to specify model components.}
\item{map_functions}{List of mapping functions used to transform bounded to unbounded parameters.}
\item{lower, upper}{Parameter bounds.}
\item{family}{The \code{family_gnlmsa} object used.}
\item{X, Z, y}{Original data matrices and response vector.}
\item{one_parameter}{Logical indicating wether the family of the object is a one parameter family.}
\item{gradient, hessian}{Gradient and Hessian at the estimated values of the parameters (free
parameters only).}
\item{nr}{The object returned from \code{gnlmsa_fit()} function.}
\item{sa}{The object returned from \code{sa_fit()} function.}
}
}
\description{
The fitting procedure consists of two stages:
\enumerate{
\item Simulated Annealing is used to explore the parameter space and identify a good candidate solution.
\item Newton–Raphson is then used to refine this candidate. If the local optimization fails, or
the Simulated Annealing solution yields a higher log-likelihood, the SA result is retained.
}
}
\details{
Fits a Generalized Non-Linear Model (GNLM) by combining global optimization via Simulated Annealing
and local refinement via Newton–Raphson. The model includes both a mean component and a dispersion
component, each of which can be specified using user-defined nonlinear predictors.
}
\examples{
\dontrun{
y <- productivity$Y
X <- productivity[, -1]
beta_start <- c(5, .5, .5)
fit <- gnlmsa(y, mean_model = cobb_douglas(X = X), beta_start, family = gnlmsa_Gamma("identity"),
              sa_control_params = sa_control(1000))
}
}
