% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{hess_mu}
\alias{hess_mu}
\title{Compute hessian with respect to parameters of mean component for Generalized Non-Linear Models}
\usage{
hess_mu(
  y,
  X,
  beta,
  mu,
  eta,
  phi,
  f_mu,
  J_mu,
  H_mu,
  mu.eta,
  mu2.eta2,
  variance,
  expected = TRUE
)
}
\arguments{
\item{y}{numerical vector of response variable.}

\item{X}{matrix containing the covariates for the mean component.}

\item{beta}{parameters of the mean component.}

\item{mu}{numerical vector containing the conditional mean of the response variable.}

\item{eta}{numerical vector containing the non linear predictor of the response variable.}

\item{phi}{numerical vector containing the conditional dispersion of the response variable.}

\item{f_mu}{function with signature \code{f(X, theta)} to compute the non linear predictor where \code{X} is the matrix containing the covariates and \code{theta} is the vector of parameters both of the mean component.}

\item{J_mu}{(optional) function to compute the jacobian of \code{f_mu()}. Should be a function with signature \code{J_mu(X, theta)} and should return a \eqn{n \times k} matrix where \code{n} is equal to \code{length(y)} and \code{k} is equal to the number of parameters of the mean component.}

\item{H_mu}{(optional) function to compute the hessian of \code{f_mu()}. Should be a function with signature \code{H_mu(X, theta)} and should return a list of \eqn{n} elements where \code{n} is equal to \code{length(y)} and each element should be the hessian of \code{f_mu()} computed at every row of \code{X}.}

\item{mu.eta}{function which computes the derivative of the mean component with respect to the non linear predictor as created by \code{\link[=make_link]{make_link()}} function.}

\item{mu2.eta2}{function which computes the second derivative of the mean component with respect to the non linear predictor as created by \code{\link[=make_link]{make_link()}} function.}

\item{variance}{function which computes the variance of the given family as a function of \code{mu} and \code{phi}.}

\item{expected}{logical indicating whether to return the expected Hessian (default \code{expected = TRUE}) or the observed (\code{expected = FALSE})}
}
\value{
a square symmetric matrix of dimension equal to the number of parameters for the mean component.
}
\description{
Compute hessian with respect to parameters of mean component for Generalized Non-Linear Models
}
\examples{
# create family object
fam_gamma <- gnlmsa_Gamma(link_mu = "identity", link_phi = "log")

# create non linear function for mean
cd <- cobb_douglas()
f_mu <- cd$f
J_mu <- cd$J
H_mu <- cd$H

# create non linear function for dispersion
lin <- Linear()
f_phi <- lin$f

# sample size
n <- 100

# covariates for the mean component
X <- cbind(rgamma(n, 1000, 10), rgamma(n, 2000, 5))

# covariates for the variance component
Z <- cbind(1, X)

# parameters for mean component
beta <- c(1, .5, .3)
eta <- f_mu(X, beta)
mu <- fam_gamma$linkinv_mu(eta)

# parameters for dispersion component
gamma <- c(1, .2, -.05)
vi <- f_phi(Z, gamma)
phi <- fam_gamma$linkinv_phi(vi)

# simulate response
y <- rgamma(n, phi, phi/mu)

hess_mu(y, X, beta, mu, eta, phi,
        f_mu, J_mu, H_mu,
        fam_gamma$mu.eta, fam_gamma$mu2.eta2, fam_gamma$variance,
        expected = TRUE)
hess_mu(y, X, beta, mu, eta, phi,
        f_mu, J_mu, H_mu,
        fam_gamma$mu.eta, fam_gamma$mu2.eta2, fam_gamma$variance,
        expected = FALSE)
}
