% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_map_function}
\alias{make_map_function}
\title{Helper function to create a function to map parameters from constrained parameter spaces to the real line}
\usage{
make_map_function(lower, upper)
}
\arguments{
\item{lower}{a numerical vector containing the lower bounds. There must be an element for each parameter of the model.}

\item{upper}{a numerical vector containing the upper bounds. There must be an element for each parameter of the model.}
}
\value{
a list with three functions:
\itemize{
\item \code{map()}: function that maps from the constrained space to the real line.
\item \code{invert()}: inverse of \code{map()} that maps back from the real line to the constrained space.
\item \code{jacobian()}: jacobian of \code{map()}.
}
}
\description{
Helper function to create a function to map parameters from constrained parameter spaces to the real line
}
\examples{

lower <- c(-Inf, 0, -1)
upper <- c(Inf, Inf, 1)

map_functions <- make_map_function(lower, upper)

map <- map_functions$map
invert <- map_functions$invert
jacobian <- map_functions$jacobian

x <- c(0, 3, 0.2)
map(x)
invert(map(x))
x - invert(map(x))
jacobian(x)
}
