% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_interval}
\alias{map_interval}
\title{Maps open intervals to real line}
\usage{
map_interval(x, lower, upper, inverse = FALSE, derivative = FALSE)
}
\arguments{
\item{x}{a numerical vector of values to map to the real line or to take bake to the open intervals.}

\item{lower}{a numerical vector containing the lower bounds of the open intervals.}

\item{upper}{a numerical vector containing the upper bounds of the open intervals.}

\item{inverse}{logical indicating whether to map from the open intervals to the real line (default, \code{inverse = FALSE}) or whether to map from the real line to the open intervals.}

\item{derivative}{logical indicating whether to return the derivative of the map function.}
}
\value{
a numerical vector containing the mapped values or the derivative.
}
\description{
Maps open intervals to real line
}
\examples{
# Create a numerical vector
x <- seq(from = 0.01, to = 0.99, length = 10)
lower <- 0
upper <- 1

# map from (0, 1) to R
y <- map_interval(x, lower, upper)

# map back from R to (0, 1)
map_interval(y, lower, upper, inverse = TRUE)

# derivative
map_interval(x, lower, upper, derivative = TRUE)
}
