% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_negative}
\alias{map_negative}
\title{Map the negative line to the real line}
\usage{
map_negative(x, upper = 0, inverse = FALSE, derivative = FALSE)
}
\arguments{
\item{x}{a numerical vector of values to map to the real line or to take back to the negative line.}

\item{upper}{value of the upper bound in order to map from \eqn{(-\infty, \texttt{upper}) \rightarrow \mathbb{R}}. \code{upper} could also be a positive value. Default is 0.}

\item{inverse}{logical indicating whether to map from the negative line to the real line (default, \code{inverse = FALSE}) or whether to map from the real line to the negative line (\code{inverse = TRUE}).}

\item{derivative}{logical indicating whether to return the derivative of the map function.}
}
\value{
a numerical vector containing the mapped values or the derivative.
}
\description{
Map the negative line to the real line
}
\examples{
# (-Inf, 0) to R
x <- -(5:1)
y <- map_negative(x)
map_negative(y, inverse = TRUE)

# (-Inf, a) to R
x <- -(6:2)
upper <- 1
y <- map_negative(x, upper = upper)
map_negative(y, inverse = TRUE, upper = upper)

# derivative
x <- -(5:1)
map_negative(x, derivative = TRUE)
}
