% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_negative}
\alias{map_negative}
\title{Create mapping functions between half-open intervals and the real line}
\usage{
map_negative()
}
\value{
A list of six functions: \code{map}, \code{invert}, \code{map_jacobian},
\code{map_hessian}, \code{invert_jacobian}, \code{invert_hessian}.
}
\description{
Returns a list of functions implementing a bijective transformation between
a half-open interval \verb{(-Inf, upper)} and the real line using a logarithmic transformation,
together with their derivatives and inverses.
}
\details{
The returned list contains:
\describe{
\item{map(x, upper)}{Maps from (-Inf, upper) to the real line: \deqn{f(x) = \log(\texttt{upper} - x)}}
\item{invert(x, upper)}{Inverse mapping: \deqn{f^{-1}(z) = \texttt{upper} - \exp(z)}}
\item{map_jacobian(x, upper)}{First derivative of the forward map: \deqn{f'(x) = -1/(\texttt{upper} - x)}}
\item{map_hessian(x, upper)}{Second derivative of the forward map: \deqn{f''(x) = -1/(\texttt{upper} - x)^2}}
\item{invert_jacobian(x, upper)}{First derivative of the inverse map: \deqn{(f^{-1})'(z) = -\exp(z)}}
\item{invert_hessian(x, upper)}{Second derivative of the inverse map: \deqn{(f^{-1})''(z) = -\exp(z)}}
}
}
\examples{
f <- map_negative()
x <- -(5:1)
f$map(x, upper = 0)
f$invert(f$map(x, upper = 0), upper = 0)
f$map_jacobian(x, upper = 0)
f$invert_jacobian(f$map(x, upper = 0), upper = 0)
f$map_hessian(x, upper = 0)
f$invert_hessian(f$map(x, upper = 0), upper = 0)
}
