% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.family_gnlmsa}
\alias{print.family_gnlmsa}
\title{Print function}
\usage{
\method{print}{family_gnlmsa}(x, ...)
}
\arguments{
\item{x}{A family_gnlmsa object}

\item{...}{further arguments passed to methods.}
}
\value{
Details about the family definition and the family object.
}
\description{
Print function
}
\examples{
gamma_fam <- gnlmsa_Gamma()
gamma_fam
}
